refactor(ui): isolate composer preview rendering
